% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-utils.R
\name{color_to_lab}
\alias{color_to_lab}
\alias{lab_to_color}
\title{Convert Colors to L\emph{a}b* and Back}
\usage{
color_to_lab(colors)

lab_to_color(lab)
}
\arguments{
\item{character}{vector or sRGB hex color codes or standard R colors that can
be interpreted by \link{col2rgb}.}
}
\description{
Colors are either standard R colors or sRGB colors encoded in hex notation
such as "#FF00E3".  These are simple wrappers around \code{grDevices} functions to
simplify translation from and to Lab space.
}
