% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-gradient-fit.R
\name{colour_ramp_equi}
\alias{colour_ramp_equi}
\title{Create a Colour Ramp Function That Produces Uniform Colours}
\usage{
colour_ramp_equi(colours, na.value, dist_fun = deltaE2000, ...,
  jnd = 1, col_to_space = color_to_lab, space_to_col = lab_to_color,
  from = 0, to = 1, eq_ctl = list(tol = 0.001, iters = 10000, quiet =
  TRUE), wesoip.tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{colours}{character a vector of colours to base the ramp on, must
have at least two colours, and the colours should be in a format understood
by \code{\link[grDevices:col2rgb]{grDevices::col2rgb}}.}

\item{na.value}{character(1L) a colour in a format understood by
\code{\link[grDevices:col2rgb]{grDevices::col2rgb}} to use to represent NA values, named to be
consistent with \code{ggplot2} conventions.}

\item{dist_fun}{a function that computes distances between colors in a three
dimensional color space.  The function should have two parameters, which
will be 3 column matrices containing the coordinates of the colours to
compute the distances between.  The function should compute the distances
between matching rows across the two matrices.}

\item{...}{arguments to pass on to \code{dist_fun}.}

\item{jnd}{numeric(1L) strictly positive, indicates the \code{dist_fun} distance
that is considered a "Just-noticeable Difference".}

\item{col_to_space}{a function that converts colors of the type accepted by
\code{\link{col2rgb}} into a 3 column matrix representing coordinates of the same
colour in the target colour space.  The default colour space is L\emph{a}b*.}

\item{space_to_col}{a function that carries out the reverse of the
that \code{col_to_space} transformation.}

\item{eq_ctl}{list, advanced usage only, arguments to pass on to
\code{\link{equalize_dists2}}.  \code{\link{equalize_dists2}} is used to make the sub-segments
we create interpolate out of the \code{colours} path equi-distance.  You may
only pass \code{tol}, \code{iters}, and \code{quiet} as the other arguments are derived
from arguments to this function.}

\item{wesoip.tol}{numeric(1L) how closely the "Whole Equal to Sum Of Its
Parts" must hold to suppress warnings about this condition not being met
when we breakdown segments of the \code{colours} path into sub-segments.  Set to
\code{Inf} to suppress warnings.  Tolerance is measured against the absolute
difference of the sum of the parts and the whole divided by the whole.}

\item{mid}{integer(1L) in \code{seq_along(colours)}, indicates which element in
\code{colours} is to be treated as the midpoint colour.}
}
\value{
a function that accepts a numeric parameter as an input, and for the
values between \code{from} and \code{to} returns a corresponding colour along the
\code{colours} path.
}
\description{
The colour ramp function converts values between 0 and 1 into colours
along the piecewise path through \code{colours} in a three dimensional colour
space, L\emph{a}b* by default.  Pairs of points that are equidistant in the input
domain should be approximately equidistant along the \code{colours} path as per
\code{dist_fun}, although if the path is not straight they are unlikely to be
equidistant in the colour space itself.
}
\details{
Distance along the path is measured as the cumulative sum of the
\code{dist_fun} distance between small segments of the path.  In order to get
the expected outcomes \code{dist_fun} should return the same distance for any
straight line segment whether it is computed as the distance from one end to
the other, or as the sum of the lengths of sub-segments that end to end
correspond to the same segment (i.e., the "Whole is Equal to the Sum Of Its
Parts").  Additionally as we move from one point to the other along the
shortest Euclidean path in the three dimensional colour space, the value
\code{dist_fun} returns should decrease, though it need not do so in proportion
to the Euclidean distance traveled.

Upon creation of the function, each segment in \code{colours} is subdivided into
enough pieces such that the distance spanned by each piece is less than the
\code{jnd}, or the "Just-Noticeable Difference".  The value-to-colour map is then
carried out by finding the segment that causes the cumulative accrued
\code{dist_fun} distance as a fraction of the total cumulative \code{dist_fun} distance
of the path to exceed the value.  The actual colour is then interpolated
linearly in that segment in the provided colour space.

The complexity of this function and the resulting colour ramp function is
designed to accommodate the CIEDE2000 colour distance function, implemented
in this package as \code{\link{deltaE2000}}, while also allowing users to specify
custom color spaces and distance functions.
}
\seealso{
\code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn}}, \code{\link[grDevices:col2rgb]{grDevices::col2rgb}},
\code{\link[grDevices:colorRamp]{grDevices::colorRamp}}.
}
