% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltae.R
\name{deltaE2000}
\alias{deltaE2000}
\alias{deltaE2000_adj}
\title{Adaptation of Bob Rudis's \code{swatch:::deltaE2000} Function.}
\usage{
deltaE2000(Labstd, Labsample, kl = 1, kc = 1, kh = 1)

deltaE2000_adj(Lab, kl = 1, kc = 1, kh = 1)
}
\arguments{
\item{Labstd}{numeric\link{, 1:3} with L, a, b coordinates}

\item{Labsample}{numeric\link{, 1:3} with L, a, b coordinates}

\item{Lab}{numeric\link{, 1:3} with L, a, b coordinates}
}
\value{
for \code{deltaE2000}, the CIEDE Delta E for corresponding rows in \code{Labstd}
and \code{Labsample}.  For \code{deltaE2000_adj}, the CIEDE Delta E between adjacent
rows of \code{Lab}.
}
\description{
Only substantive change is to remove the colorspace dependency, directly
accepting Lab matrices instead of the S4 LAB objects from colorspace.
}
\details{
CIEDE2000 implementation as per:

Gaurav Sharma & Maynard P Baalthazar
}
