% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltae.R
\name{deltaE2000_1}
\alias{deltaE2000_1}
\title{Adaptation of Bob Rudis's \code{swatch:::deltaE2000} Function.}
\usage{
deltaE2000_1(Labstd, Labsample, kl = 1, kc = 1, kh = 1)
}
\arguments{
\item{Labstd}{numeric\link{, 1:3} with L, a, b coordinates}

\item{Labsample}{numeric\link{, 1:3} with L, a, b coordinates}
}
\description{
Only substantive change is to remove the colorspace dependency, directly
accepting Lab matrices instead of the S4 LAB objects from colorspace.
}
\details{
CIEDE2000 implementation as per:

Gaurav Sharma & Maynard P Baalthazar
}
