% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.car.data.R
\docType{data}
\name{geom.car.data}
\alias{geom.car.data}
\title{Car Coordinates}
\format{A data frame with 8 rows and 6 variables
\describe{
\item{vehicle}{vehicle id}
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{length}{vehicle length}
\item{width}{vehicle width}
\item{label}{vehicle description}
}}
\usage{
geom.car.data
}
\description{
A data frame with car coordinates and sizes for use in examples.
}
\keyword{datasets}
