% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto.R, R/geom-waterfall.R,
%   R/position-waterfall.R
\docType{data}
\name{ggbg-ggproto}
\alias{ggbg-ggproto}
\alias{GeomWaterfall}
\alias{PositionWaterfall}
\title{ggbg ggproto Objects}
\format{An object of class \code{GeomWaterfall} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
GeomWaterfall
}
\description{
These are all the \code{ggproto} objects implemented by \code{ggbg}.
}
\keyword{datasets}
