% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-utils.R
\name{grob_palette}
\alias{grob_palette}
\title{Display a Palette in a Table}
\usage{
grob_palette(colors, name = deparse(substitute(colors)[[1]]))
}
\description{
Inspired by pals::pal.zcurve, although able to do rectangular (half square)
tables as well.
}
