% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-waterfall.R
\name{position_waterfall}
\alias{position_waterfall}
\title{Stack Chart Elements on Cumulative Value}
\usage{
position_waterfall(width = NULL, preserve = c("total", "single"),
  reverse = FALSE, dodge = TRUE, vjust = getOption("ggbg.vjust"),
  vjust.mode = getOption("ggbg.vjust.mode"),
  signif = getOption("ggbg.signif"), y.start = 0)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}

\item{dodge}{TRUE (default) or FALSE, controls how to resolve
groups that overlap on the \code{x} axis.  The default is to dodge them
to form mini-waterfalls within each \code{x} value, but you can chose to stack
them instead by setting \code{dodge=FALSE}.  Negative and positive values are
segregated prior to stacking so they do not overlap.  Interpreting
waterfall charts with stacked sub-groups is difficult when they contain
negative values, so we recommend you use the default setting instead.
Observations within a group that have the same \code{x} value are always
stacked, so if you have both positive and negative values for any given \code{x}
value you may want to consider segregating the positives and negatives in
different groups.}

\item{vjust}{like the \code{vjust} parameter for \code{\link[ggplot2:position_stack]{ggplot2::position_stack}},
except that by default the direction of justification follows the direction
of the bar (see \code{vjust.mode}), and the default value is \code{0.5} instead of
\code{1}.  This only has an effect on geoms with positions like text, points, or
lines.  The default setting places elements midway through the height of
the corresponding waterfall step.  The default value is convenient for
labeling \code{geom_col} waterfalls.  Use \code{1} to position at the "end" of each
waterfall step.  This is different to the \code{vjust} for geoms like
\code{geom_text} where \code{vjust=1} shift the text down, but it is consistent with
what \code{\link[gggplot2:position_stack]{gggplot2::position_stack}} does.}

\item{vjust.mode}{character(1L), one of "end" (default), or "top" where "top"
results in the same behavior as in \code{\link[ggplot2:position_stack]{ggplot2::position_stack}}.  "end"
means the justification is relative to the "end" of the waterfall bar.  So
if a waterfall bar is heading down (i.e. negative \code{y} value), the "end" is
at the bottom.  If it heading up (i.e. positive \code{y} value), the "end" is at
the top.  For positive \code{y} values "end" and "top" do the same thing.}

\item{signif}{integer(1L) between 1 and 22, defaults to 11, corresponds to
the \code{digits} parameter for \code{\link{signif}} and is used to reduce the precision
of numeric \code{x} aesthetic values so that stacking is not foiled by double
precision imprecision.}

\item{y.start}{numeric(1L), defaults to 0, will be starting point for the
cumulative sum of \code{y} values.  This could be useful if you want to combine
waterfalls with other layers and need the waterfall to start at a specific
value.}

\item{dodge}{TRUE or FALSE (default), whether to dodge waterfall bars when
there are multiple bars for a single x value.}
}
\description{
A waterfall chart is a bar chart where each segment starts where the prior
segment left off.  This is similar to a stacked bar chart, except that
the stacking does not reset across \code{x} values.  It is the visualization of a
cumulative sum.  Another similar type of chart is the candlestick plot,
except those have "whiskers", and typically require you to manually specify
the \code{ymin} and \code{ymax} values.
}
\details{
\code{position_waterfall} creates waterfall charts when it is applied to
\code{geom_col} or \code{geom_bar}.  You can apply it to any geom, so long as the
geom specifies a \code{y} aesthetic, and either an \code{x} aesthetic, or both
\code{xmin} and \code{xmax} aesthetics.  It may not make sense to apply
\code{position_waterfall} to arbitrary geoms, particularly those that represent
single graphical elements with multiple x/y coordinates such as
\code{geom_polygon}.

Since stat layers are computed prior to position adjustments, you can also
use \code{position_waterfall} with stats (e.g \code{stat_bin}, see examples).

We also implement a \link[=ggbg-ggproto]{StatWaterfall} \code{ggproto} object that
can be accessed within \code{geom_*} calls by specifying \code{stat='waterfall'}.
Unlike typical stat \code{ggproto} objects, this one does not have a layer
instantiation function (i.e. \code{stat_waterfall} does not exist).  The sole
purpose of the stat is to compute the \code{ycum} aesthetic that can then be used
by the \code{geom} layer (see the labeling examples).
}
\section{Stacking}{


The stacking is always computed on the \code{y} aesthetic.  The order of the
stacking is determined by the \code{x} aesthetic.  The actual position of the
objects are also affected by \code{vjust}, and you may need to change the value of
\code{vjust} if you are using \code{position_waterfall} with geoms other than columns.
For example, for dimensionless elements such as \code{geom_point} and \code{geom_text},
the default \code{vjust} of 0.5 leads to alignment at the midpoint between
previous and subsequent values in the cumulative sequence.

If only \code{xmin} and \code{xmax} aesthetics are present the \code{x} value will be
inferred as the midpoint of those two.
}

\section{Dodging}{


Unlike most \code{position_*} adjustments, \code{position_waterfall} adjust positions
across different \code{x} values.  However, we still need to resolve \code{x}
value overlaps.  The default approach is to apply the same type of adjustment
across groups within any given \code{x} value.  This stacks and dodges elements.

Dodging involves changing the \code{width} of the geom and also shifting the
\code{geom} horizontally.  Geom width adjustments will always be made based on the
\code{xmin}/\code{xmax}/\code{width} aesthetics.  The shifting itself can be controlled
separately with \code{position_waterfall(width=...)}.  That parameter should
really be called \code{dodge.width} to avoid confusion with the geom \code{width}, but
we left it as \code{width} for consistency with \code{\link{position_dodge}}.

You you can turn off dodging within \code{x} values by setting
\code{position_waterfall(dodge=FALSE)} which will result in stacking within each
\code{x} value.
}

\examples{
## These examples are best run via `example(position_waterfall)`
library(ggplot2)
dat <- data.frame(x=3:1, y=1:3)
p1 <- ggplot(dat, aes(x=x, y=y)) + geom_col(position='waterfall')

## Add text or labels; defaults to middle waterfall position
## which can be modified with `vjust`
p1 + geom_label(aes(label=x), position='waterfall')

## We can also add the cumulative running to the top of
## the bars with `stat='waterfall'` and position adjustments
p1 + geom_label(aes(label=x), position='waterfall') +
 geom_label(
   stat="waterfall",        # adds `ycum` computed variable
   aes(label=stat(ycum)),   # which we can use for label
   position=position_waterfall(vjust=1), # text to end of column
   vjust=0,                              # tweak so it's on top
)
## A poor person's candlestick chart:
dat.r.walk <- data.frame(x=1:20, y=rnorm(20))
ggplot(dat.r.walk, aes(x=x, y=y, fill=y > 0)) +
  geom_col(position='waterfall')

## We can use arbitrary geoms
ggplot(dat, aes(x=x, y=y)) +
  geom_point() +
  geom_point(position='waterfall', color='blue') + # default vjust=0.5
  geom_point(position=position_waterfall(vjust=1), color='red')

## Or stats; here we turn a histogram into an ecdf plot
dat.norm <- data.frame(x=rnorm(1000))
ggplot(dat.norm, aes(x=x)) + geom_histogram(position='waterfall')
ggplot(dat.norm, aes(x=x)) + stat_bin(position='waterfall')

## Data with groups
dat3 <- data.frame(
  x=c(3, 2, 2, 2, 1, 1), y=c(-3, 1, 4, -6, -1, 10),
  grp=rep(c("A", "B", "C"), lenght.out=6)
)
p2 <- ggplot(dat3, aes(x=x, y=y, fill=grp))
p2 + geom_col(position="waterfall")

## Equal width columns
p2 + geom_col(position=position_waterfall(preserve='single'))

## Stacking groups is possible, bug hard to interpret when
## negative values present
p2 + geom_col(position=position_waterfall(dodge=FALSE))
}
