% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-gradient-fit.R
\name{colour_ramp_equi}
\alias{colour_ramp_equi}
\title{Create a Colour Ramp Function That Produces Uniform Colours}
\usage{
colour_ramp_equi(colours, na.value, dist_fun = deltaE2000, ...,
  jnd = 0.75, col_to_space = color_to_lab,
  space_to_col = lab_to_color, from = 0, to = 1, eq_ctl = list(tol
  = 0.001, iters = 10000, quiet = TRUE))
}
\arguments{
\item{colours}{character a vector of colours to base the ramp on, must
have at least two colours, and the colours should be in a format understood
by \code{\link[grDevices:col2rgb]{grDevices::col2rgb}}.}

\item{na.value}{character(1L) a colour in a format understood by
\code{\link[grDevices:col2rgb]{grDevices::col2rgb}} to use to represent NA values, named to be
consistent with \code{ggplot2} conventions.}

\item{dist_fun}{a function that computes distances between colors in a three
dimensional colour space.  The function will be provided with two
three-column matrices representing coordinates in the three dimensional
colour space.  The function should compute the distances between matching
rows across the two matrices.  It is expected that the function will
produce monotonically decreasing values as we move along the shortest
Euclidean path between two points in the three dimensional colour space.
The function may accept additional arguments (see \code{...}).}

\item{...}{arguments to pass on to \code{dist_fun}.}

\item{jnd}{numeric(1L) strictly positive, indicates the \code{dist_fun} distance
that is considered a "Just-noticeable Difference".  The path along
\code{colours} is broken up into sub-segments intended to be shorter than \code{jnd}.
If \code{dist_fun} is not guaranteed to return the same length for a segment
whether it is computed in one pass or as the sum of component
sub-segments, then it is not guaranteed that the sub-segments are not
guaranteed to be all shorter than \code{jnd}.}

\item{col_to_space}{a function that converts colors of the type accepted by
\code{\link{col2rgb}} into a 3 column matrix representing coordinates of the same
colour in the target colour space.  The default colour space is L\emph{a}b*.}

\item{space_to_col}{a function that carries out the reverse of the
that \code{col_to_space} transformation.}

\item{eq_ctl}{list, advanced usage only, arguments to pass on to
\code{\link{equalize_dists2}}.  \code{\link{equalize_dists2}} is used to make the sub-segments
we create interpolate out of the \code{colours} path equi-distance.  You may
only pass \code{tol}, \code{iters}, and \code{quiet} as the other arguments are derived
from arguments to this function.}

\item{mid}{integer(1L) in \code{seq_along(colours)}, indicates which element in
\code{colours} is to be treated as the midpoint colour.}
}
\value{
a function that accepts a numeric parameter as an input, and for the
values between \code{from} and \code{to} returns a corresponding colour along the
\code{colours} path.
}
\description{
The colour ramp function converts values between 0 and 1 into colours
along the piecewise path through \code{colours} in a three dimensional colour
space, L\emph{a}b* by default.  Pairs of points that are equidistant in the input
domain should be approximately equidistant along the \code{colours} path as per
\code{dist_fun}, although if the path is not straight they are unlikely to be
equidistant in the colour space itself.  How distance is computed is
specified by the \code{dist_fun} function.  The default \code{\link{deltaE2000}} is a
non-Euclidian distance measure.
}
\details{
Upon creation of the function, each segment in \code{colours} is subdivided into
enough pieces such that the distance spanned by each piece is less than the
\code{jnd}, or the "Just-Noticeable Difference".  The value-to-colour map is then
carried out by finding the segment that causes the cumulative accrued
\code{dist_fun} distance as a fraction of the total cumulative \code{dist_fun} distance
of the path to exceed the value.  The actual colour is then interpolated
linearly in that segment in the provided colour space.

To the extent \code{dist_fun} returns different results when computing distance of
a full segment and the sum of the distance of the segments that it is
composed of, the sub-segments may end up longer than \code{jnd}.

The complexity of this function and the resulting colour ramp function is
designed to accommodate the CIEDE2000 colour distance function, implemented
in this package as \code{\link{deltaE2000}}, while also allowing users to specify
custom colour spaces and distance functions.
}
\seealso{
\code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn}}, \code{\link[grDevices:col2rgb]{grDevices::col2rgb}},
\code{\link[grDevices:colorRamp]{grDevices::colorRamp}}.
}
