% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto.R, R/position-waterfall.R,
%   R/stat-waterfall.R
\docType{data}
\name{ggbg-ggproto}
\alias{ggbg-ggproto}
\alias{PositionWaterfall}
\alias{StatWaterfall}
\title{ggbg ggproto Objects}
\format{An object of class \code{StatWaterfall} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatWaterfall
}
\description{
These are all the \code{ggproto} objects implemented by \code{ggbg}.

\code{PositionWaterfall} is the \code{ggproto} object used to generate the position
adjustments that correspond to \link{position_waterfall}.

\code{StatWaterfall} is intended for use only via the \code{stat} parameter to other
layers (e.g. \code{geom_...(stat='waterfall')}).  It will add a \code{ycum} aesthetic
to the data which can then be used by the layer.  Unlike other \code{Stat}
objects, it does not exist as a stand-alone layer.  Its main purpose is to
generate the text for cumulative labels.  See \code{\link{position_waterfall}} for
usage examples.
}
\seealso{
\code{\link{position_waterfall}} for usage examples.
}
\keyword{datasets}
