% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-waterfall.R
\name{position_waterfall}
\alias{position_waterfall}
\title{Stack objects in a Waterfall}
\usage{
position_waterfall(width = NULL, preserve = c("total", "single"),
  reverse = FALSE, dodge = FALSE)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}

\item{dodge}{TRUE or FALSE (default), whether to dodge waterfall bars when
there are multiple bars for a single x value.}
}
\description{
Stack objects in a Waterfall
}
